feat: enhanced literals — hex/binary ints, U/L/UL suffixes, digit separators, f/d suffixes, char escapes (CVL1900-1910)

+++ src/Cvolo.Analysis/Passes/DeclarationPass.cs
@@ -2513,7 +2513,7 @@ public sealed class DeclarationPass(BindingContext context)
 		switch (expr)
 		{
 			case IntegerLiteralExpressionSyntax intLit:
-				return intLit.Value;
+				return (long)intLit.Value;
 			case UnaryExpressionSyntax { Operator: "-" } unary:
 				var operand = EvaluateEnumConstant(unary.Operand, variantLookup);
 				return operand is null ? null : -operand.Value;
